feat: generate square dalitz
