+++ tf_pwa/config_loader/sample.py
@@ -4,7 +4,7 @@ from tf_pwa.amp.core import get_particle_model_name
 from tf_pwa.cal_angle import cal_angle_from_momentum
 from tf_pwa.config import get_config
 from tf_pwa.data import data_mask, data_merge, data_shape
-from tf_pwa.generator.generator import BaseGenerator, GenTest
+from tf_pwa.generator.generator import BaseGenerator, GenTest, multi_sampling
 from tf_pwa.particle import BaseParticle
 from tf_pwa.phasespace import ChainGenerator  # as generate_phsp_o
 from tf_pwa.tensorflow_wrapper import tf
@@ -187,57 +187,6 @@ def generate_toy_p(
         config.max_amplitude = max_weight
 
     return ret
-
-
-def multi_sampling(
-    phsp, amp, N, max_N=200000, force=True, max_weight=None, importance_f=None
-):
-    a = GenTest(max_N)
-    all_data = []
-
-    for i in a.generate(N):
-        data, new_max_weight = single_sampling2(
-            phsp, amp, i, max_weight, importance_f
-        )
-        if max_weight is None:
-            max_weight = new_max_weight * 1.1
-        if new_max_weight > max_weight and len(all_data) > 0:
-            tmp = data_merge(*all_data)
-            rnd = tf.random.uniform((data_shape(tmp),), dtype=max_weight.dtype)
-            cut = (
-                rnd * new_max_weight / max_weight < 1.0
-            )  # .max_amplitude < 1.0
-            max_weight = new_max_weight * 1.05
-            tmp = data_mask(tmp, cut)
-            all_data = [tmp]
-            a.set_gen(data_shape(tmp))
-        a.add_gen(data_shape(data))
-        # print(a.eff, a.N_gen, max_weight)
-        all_data.append(data)
-
-    ret = data_merge(*all_data)
-
-    if force:
-        cut = tf.range(data_shape(ret)) < N
-        ret = data_mask(ret, cut)
-
-    status = (a, max_weight)
-
-    return ret, status
-
-
-def single_sampling2(phsp, amp, N, max_weight=None, importance_f=None):
-    data = phsp(N)
-    weight = amp(data)
-    if importance_f is not None:
-        weight = weight / importance_f(data)
-    new_max_weight = tf.reduce_max(weight)
-    if max_weight is None or max_weight < new_max_weight:
-        max_weight = new_max_weight * 1.01
-    rnd = tf.random.uniform(weight.shape, dtype=weight.dtype)
-    cut = rnd * max_weight < weight
-    data = data_mask(data, cut)
-    return data, max_weight
 
 
 class AfterGenerator(BaseGenerator):
